Fix LFS directory structure of the project
